Fix bug where old pass is still displayed.
We needed to split loading into two dosync blocks so that pass detection
happens before we go changing the UI based on the newly loaded passes.
When everything was in one dosync block, we were looking at the *old*
set of passes to decide things like which one to select.